+++ src/main/java/online/pictz/api/vote/service/memory/VoteInMemoryService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import online.pictz.api.vote.dto.VoteRequest;
 import online.pictz.api.vote.service.VoteService;
+import online.pictz.api.vote.service.memory.atmoic.AtomicChoiceStorage;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +13,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class VoteInMemoryService implements VoteService {
 
-    private final InMemoryChoiceStorage choiceStorage;
+    private final AtomicChoiceStorage choiceStorage;
 
     @Override
     public void voteBulk(List<VoteRequest> voteRequests) {
